Adding codon library, slightly trimmed.
Codon is a library I worked on a few years ago to try out some opinionated
messaging ideas in under Java. Returning to it to try out some messaging
stuff under this.

It's effectively a wrapper around Camel with some extra glitz for prettier
application code and for handling subscriptions, discovery, etc..